+++ src/Controller/VirtualAliasController.php
@@ -31,6 +31,9 @@ class VirtualAliasController
         $params = $request->getParams();
 
         try {
+            if(empty($params['sourceId']) || empty($params['destination'])) {
+                throw new Exception('Email de origem ou alias não informado');
+            }
             $alias = $this->valias->create($params['sourceId'], $params['destination']);
             return $this->toJson($response, 200, 'ok', $alias);
         } catch(Exception $ex) {
@@ -42,6 +45,9 @@ class VirtualAliasController
     {
         $params = $request->getParams();
         try {
+            if(empty($params['aliases'])) {
+                throw new Exception('Nenhum alias foi informado');
+            }
             $this->valias->delete($params['aliases']);
             return $this->toJson($response, 204);
         } catch(Exception $ex) {
